+++ src/lib.rs
@@ -37,10 +37,6 @@ extern crate quick_error;
 #[macro_use]
 extern crate serde_derive;
 #[macro_use]
-extern crate slog_derive;
-#[macro_use]
-extern crate derive_more;
-#[macro_use]
 extern crate more_asserts;
 #[macro_use]
 extern crate vlog;
